kafferepet/requirements.txt
feedparser
requests
openai-whisper